Ensure mount item is properly set to bound after update
Summary:
When calling updateMountItemsIfNeeded in Rendercore MountState, the call to updateExtensions will call unbind and bind on the component. At the end of this sequence, we know that a component is bound. However, there is no reflection of this in MountItem, and so we risk going out of sync and having MountState think the item is unbound, and bind it again.
We solve this by ensuring the MountItem is set to bound (true) after calling updateExtensions.

Reviewed By: adityasharat

Differential Revision: D31612799

fbshipit-source-id: dbca8826cb07665096295b8d9f189b76d54fbc2c