Refactors AP with information of LayoutInterStageProps Annotations
Summary:
Threads through the upcoming `LayoutInterStageInputParamModel`, and the annotations for the `LayoutInterStagePropsContainer` in litho's AP extractor methods. This is change set passes an empty collection of Annotations which can be used as layout interstage prop inputs in spec methods.

The new `LayoutInterStageInputParamModel` will allow the AP to switch the target container to read and write the inter stage props.

Reviewed By: astreet

Differential Revision: D31271604

fbshipit-source-id: b4ec7b6587a9c3b8000021486b4da8b9a9fd65ef