Makes changes to AP to pass the inter stage props
Summary:
The current AP generates a method in the component body to to get the `InterStagePropsContainer`. Now the inter stage props are being passed as a method argument this diff does some makes in the AP to use that method argument instead.

1. Refactors the getter from `<base-class>.getInterStagePropsContainer(<spec-context>)` to `<base-class>.getInterStagePropsContainer(<spec-context>, InterStagePropsContainer)`
2. Captures the name of the method argument for the inter stage props container (since it is never directly accessed by the spec methods, a place holder name is generated (`_1`, `_2`, and so on).
3. The captured name is passed to `getImplAccessor()` utility method to generate read and write statements.

This leads to an interim state where the follow code statements may get generated.

```
InterStatePropsContainer _interStageProps = null;
```

Reviewed By: astreet

Differential Revision: D31172280

fbshipit-source-id: baca6aac9563ffbd3e9e3c44a170b1efbad91415