Don't unbind the LithoView in HorizontalScroll/VerticalScroll unmount
Summary:
Since there's no reciprocal bind in mount, this can cause issues where the nested tree is left in an unbound state.

The alternative is to add a bind call in mount - however, the status quo for almost all MountSpecs hosting nested trees is to not propagate down bind/unbind calls (this is generally done automatically from onAttach/DetachFromWindow). So to keep there from being as much surface area for additional issues, we're just going to make this consistent with the other MountSpecs.

Reviewed By: adityasharat

Differential Revision: D31861330

fbshipit-source-id: f104fb2cd5e4bd2a13d1af8471b3f5f5e5c9a4a2