Fix rest of tests after shipping renderunit id map
Summary:
These two tests actually broke because we can do a *better* job of referencing components to mount content across re-renders than before!

For the TreeDiffingTest, we can now compare the correct mount contents via diffnodes and know they haven't changed.

For the DynamicPropsTest, we are no longer trying to use the view that showed the first child with dynamic props to render the second child when that first child goes away. This means the test has had to be rewritten a bit to test that we correctly unset dynamic props when they don't exist anymore.

Differential Revision: D31957320

fbshipit-source-id: 9b6382a0d139a67d8c5697d5ee4f33c3a6242cca