Don't store Builder on CollectionLayout
Summary: Not a huge deal, but happened on this while changing other code. This can create a subtle bug where we have mutated the LinearRecyclerConfiguration builder stored on the layout object and then reuse it to build a different configuration. I don't think this is possible atm but it's not crazy that it could happen.

Reviewed By: andrewpmsmith

Differential Revision: D31959684

fbshipit-source-id: 8eff4c7e04fb37c01c2462c7310a1e6ecd494ecb