Fix boundary condition for scrolling down
Summary: IncrementalMountExtension had an issue with the scrolling down boundary condition wherein if an item is exactly out-of-screen (item bottom = visible bounds top), it would be unmounted in the 1st loop, and then immediately mounted again in the 2nd loop. Since the item is fully offscreen, the correct behaviour is for the item to be unmounted and not mounted again. Since both condition were checking for the "equals" condition in the boundary, the 2nd loop that mounts item again no longer checks for equality.

Reviewed By: adityasharat

Differential Revision: D32099189

fbshipit-source-id: ca01768bd394185fae4bd7b6616c66a89e2dff54