Init sticky header in layout when component tree is null and dx,dy==0
Summary:
The issue is that both switching sections or scrolling to top cause the sticky header component tree to be set to null. However, we only want to re-init the sticky header when switching sections. If sticky header is allowed to be initialized while programmatically scrolling to top, `SectionsRecyclerView#setStickyComponent` detaches the component tree from the component, causing it to disappear. The fix is to add a check for no scrolling action with `dx==0 && dy==0`.

Renamed `initStickyHeaderWhenComponentTreeIsNull` to `initStickyHeaderInLayoutWhenComponentTreeIsNull`

Remove `initStickyHeaderOnScroll`

Reviewed By: astreet, Ambroos

Differential Revision: D32074895

fbshipit-source-id: fa4c2af5bc4c2e1b9eca8407275864a01e8f9684