Include component type and frequency in generated id
Summary:
We want id's to be as stable as possible across re-creations.

In the following example we need to generate an id for each child. Currently, ids are generated in order, so adding new children in the middle will mean ids get applied to different children. This diff updates the id generation to take component type into account.

```
override fun ComponentScope.render(): Component {
  val showLabel = useState {false}
  return Collection {
    if (showLabel.value) { child(Text("Hello") }
    child(child(MyComponent())) // <-- id should be independent of showLabel.value
  }
}
```

Reviewed By: astreet

Differential Revision: D32170640

fbshipit-source-id: 68d9d1693a1a067b64acefdac77c7493e1323af0