Preload Litho classes during IG startup
Summary:
putting this up to start a conversation about what is the recommended way to ensure we preload Litho classes during startup and that they end up in the IG betamap.
I've seen this approach done for fb4a Composer https://fb.workplace.com/notes/335701967538203, though it seems like we'd need to enable Loom tracing to understand what classes to preload.
Is there a better way of doing this?

Reviewed By: astreet

Differential Revision: D32100218

fbshipit-source-id: ffc99e672737411612b767392eb81e7137eb6743