Add hasProps + Matcher
Summary:
Value equality is useful if equals is implemented, but it isn't so useful if you want to test that it's just an instance of x, y, z, etc.  (Or if you even want to do some additional custom matching.)  In the specific case I hit, I wanted to just check that a property was a specific instance of an interface (though I happened to have a singleton, so the value check was fine).

This exposes passing a matcher to use to check against a prop, which provides additionally flexibility should folks not want to just do equals checks

Differential Revision: D32305022

fbshipit-source-id: 0b2728073d1face6a3a2127a81476357d8b30ae2