Add Text Containing Assertions
Summary:
If you're adding a spannable string in a component, you're often combining multiple strings together (often times when a unicode space character).  This adds the ability to check that the component renderers text that contains the given text.

This essentially is a nicer wrapper around "matching" which adds the necessary regex characters to make this work.

Differential Revision: D32305242

fbshipit-source-id: 786bacb8d9eb797fc1e099e924519b175122a8ad