Move HostComponent check to Binder-specific shouldUpdate
Summary: Previously, the early return in should-update sequence for HostComponents was being used by both the binder and the AttributesExtension. This meant that the AttributesExtension was not triggering an update for HostComponents when it should have been. The early return for HostComponent should only be relevant for the mount binder to avoid unmount / mount - and it should not prevent the attributes from being applied.

Reviewed By: astreet

Differential Revision: D32429732

fbshipit-source-id: c4645ca9299d805d4a14d11f1e86b88a6e0e46ae