Clean up DynamicPropsManager entry on unbind
Summary:
We never clean up the `mAffectingDynamicValues` entries of DynamicPropManager, even though the entries are always reset in bind. This can lead to accumulating large objects that the Component might be holding on to, especially for non-stateless Components.
In IG reels every item in the feed list uses a dynamic value, and in local heap dumps there are lots of Component references adding up in DynamicPropsManager#mAffectingDynamicValues that hold on to ComponentTree objects indirectly

Differential Revision: D32718536

fbshipit-source-id: 888a1fcc0ab07210ab8112048515f8f62fca5c12