Make ComponentTree#setRootAndSizeSpec() default to async
Summary:
Currently ComponentTree provides a default version of setRootAndSizeSpec which triggers a sync layout calculation, as well as an async version which calculates layout in the background but needs to be invoked explicitly through setRootAndSizeSpecAsync.
We want the default version to be async, so that developers only resort to the sync version intentionally.
The new APIs are:
setRootAndSizeSpec for async layout calculation
setRootAndSizeSpecSync for sync layout calculation

This diff handles changing setRootAndSizeSpec -> setRootAndSizeSpecSync

Reviewed By: mihaelao

Differential Revision: D32560685

fbshipit-source-id: 33c101028c9abc36bc8ac56f41a97c7215081a13