ComponentsRule -> LithoViewRule
Summary:
With the new wrapper around LithoViewRule I want to make it easy to call:
assertThat(wrapperClass)....

Removed ComponentsRule and changed it to LithoViewRule.
My first plan was to move ComponentsRule to jassert package and unblock the Wrapper, but it was too tricky with dependencies.
So instead i fastmod all of the usage to use LithoViewRule instead...

(I hope regex wont fail me but i will wait for all of the tests now)

Reviewed By: astreet

Differential Revision: D32830888

fbshipit-source-id: 1ff60d79b3bc4edae32d8a635d043569ef095095