Use last transition input to determine old root mount item count
Summary:
Previously, we were incorrectly using the mount item count in the host view directly. This is flawed, since mount items might've been removed from the host via other flows (IME). To solve this, we instead use the last-input to assess the root's mount item count similar to how we already assess the new input for the incoming frame.
Added some slight refactors to accommodate for this change.

Reviewed By: adityasharat

Differential Revision: D32954588

fbshipit-source-id: c0d91c07418077caabd8ae9fbfcc558d47c409af