Build log string only when mEnsureParentMount is false.
Summary: RenderCore MountState includes logic to collect loggable information while mounting in a StringBuilder to be used in an exception when we attempt to mount an item whose parent isn't mounted. However, we will only ever throw an exception in this case when mEnsureParentMounted is false. When true, we will just mount the parent directly. To avoid unnecessary performance overhead and memory allocation, we now only create this StringBuilder and append content to it when mEnsureParentMounted is false.

Reviewed By: adityasharat

Differential Revision: D33126724

fbshipit-source-id: 84d1b88583ae4114abcff2a307b1368aca022006