Convert LayoutState's mRenderUnitIdsWhichHostRenderTrees ArraySet to a HashSet
Summary: When comparing performance of an ArraySet with a HashSet side-by-side, HashSet.contains seems to run 2-3 times faster compared to ArraySet.contains. In FB4A, the mRenderUnitIdsWhichHostRenderTrees is used to check if a certain ID is contained a lot (e.g., navigating to the Home tab in FB4A adds about 2k calls to this). Therefore, to optimise performance here, this set is now a HashSet.

Differential Revision: D33160036

fbshipit-source-id: feae3c3574c8a500892f874a83c2a4331be6da25