Remove InternalNodeFactory from ComponentTestHelper
Summary:
Added a static resolve method which copies the implementation of Row.resolve and Column.resolve but wraps the child in a node that gets shallow-resolved. This emulates the behaviour that TestDefaultInternalNode was achieving.

Added a copy of Layout.create which uses the new resolve method and used that to shallow-resolve the tree.

Reviewed By: adityasharat

Differential Revision: D33257983

fbshipit-source-id: fe4fd4e6c0525af96cbd145d59442bac8fe0927c