Add Handle Documentation
Summary:
This diff mainly adds (basic) documentation on how to use/create handles.  In addition it also ports over a Java example to Kotlin.

Facebok:
Should this also go in the cheat sheet?  It felt a bit weird to add "Handle" --> "handle" in the API, but it might be worthwhile given the support post that spawned this.

~~WARNING: The given example isn't working for me -- likely because the trigger is `null`, so we never request focus (or at least, that is what it looks like given the bytecode isn't matching between running the sample app and adding breakpoints into the generated TextInput code).  Wanted to get feedback on what might be good to add as an example instead (if this is expected) -- one thing I considered was triggering a tooltip via a handle.~~

Reviewed By: strulovich

Differential Revision: D33357634

fbshipit-source-id: 2a0319eef9f9f859637325e6ca19f1db7ab134f6