Cache value of shouldUpdate to avoid duplicate calls while updating
Summary:
shouldUpdate gets called from both LithoRenderUnit & LithoAttributesExtension during update sequence and in immediate succession. Since shouldUpdate may be a heavy operation, this should be optimised.

To solve this, RenderUnit has been given 2 empty-impl protected methods: "onStartUpdateRenderUnit" and "onEndUpdateRenderUnit". LithoRenderUnit will override these to let it know that it can begin (and stop) caching the result of shouldUpdate. When the result is cached, it is returned without needing to calculate again.

Reviewed By: adityasharat

Differential Revision: D33623187

fbshipit-source-id: 5cb8de789fe244fe747f49d4b03f276158bd7e33