Ignore padding in RenderCore MountState when used from a LithoView
Summary: In Litho, we always ignore padding when applying bounds. In RCMS we always pass bounds, which forces us to trigger an "instanceof" check. Normally, this instanceof check is very fast, however since this API is called per mount-item, per frame, this can add up to quite a lot. To reflect this behaviour in RCMS, LithoRenderUnits will now always provide a null padding.

Reviewed By: adityasharat

Differential Revision: D33633321

fbshipit-source-id: ba46a08343d0f973b42530774139831c812af267