Optimise finding index of RenderTreeNode ID in a RenderTree
Summary:
In RenderCore MountState, we look for the index of a node in a tree in many flows:
* notifyMount
* unmountOrMoveOldItems
* maybeEnsureParentIsMounted (on each mountRenderUnit)

Each time this happened, the code iterates the RenderTree's flat-list until such an index is found, running in O(n). This is made worse by the fact that in some cases, this is done for each item in the tree, making these methods run in O(n^2).
To optimise this, when the RenderTree is created, it generates an ID-to-index map from the provided flat list, thus finding an index of a render-tree-node ID is O(1)

Reviewed By: astreet

Differential Revision: D33741768

fbshipit-source-id: 03637881ad7f2526b0b7a90144741bddfe70a6b7