Ensure LithoRenderUnit returns the component Class when getting the description
Summary: The getDescription method is used in RenderCore MountState when beginning a systrace section of mounting a render-unit. In this case, getDescription will always just return the LithoRenderUnit class, which is not helpful. To help make the systrace section clearer, LithoRenderUnit will now override getDescription to return the component's class instead

Reviewed By: astreet

Differential Revision: D33742374

fbshipit-source-id: e90a575887164c1e632afd89e2ccdcbe167f55f7