CollectionContainerScope extends ResourcesScope
Summary:
`CollectionContainerScope` now extends ResourcesScope.

This is to assist in breaking collections into separate functions without having to
1. pass both CollectionContainerScope and ComponentScope (messy) or...
2. Use SubCollections which has it's own gotchas

  fun CollectionContainerScope.addStuff() {
    // Easily access functions/vars scoped on
    // CollectionContainerScope and ResourcesScope
    // e.g. child() and Text()
    child(Text("Hello World"))
  }

  Collection {
    addStuff()
  }

NB: Conversion of product code to this pattern will follow in stacked diffs.

EDIT: This change also required me to make `FlexboxContainerScope` extend `ResourcesScope`. This is to avoid having to use this@ syntax to specify the explicit receiver.

Reviewed By: astreet

Differential Revision: D33685488

fbshipit-source-id: 398489755650ca7c10abb20b50c8fbdd28b8599a