Treat null common props as equivalent
Summary:
We were not treating null common props as equivalent.

This meant that we would fail the `isComponentEquivalent` check with two components that were equivalent, but had null common props.

Note that it's fine/expected for common props to be null, it's only assigned when a common prop is added.

Reviewed By: astreet

Differential Revision: D33918384

fbshipit-source-id: b959ee2f6b0271324ee0fdda373c734f51cf690d