Separate binding from the concept of LazyTree
Summary:
The fact that Bind has been so far implemented in a LazyTree implementation is not great as LazyTree is a RenderCore interface and it feels like Bind should be independent. The only reason why this works is that BloksModel also implements Node. That's supposed to be an optimisation though and one that we might want to really evaluate again moving forward.

Interestingly this also solves a problem with bind being called more than once for Litho trees measuring more than once.

Reviewed By: rooju

Differential Revision: D33818565

fbshipit-source-id: dc07187840c60f2bbec10f2191a125cb26d054fd