Add experiment to recycle ComponentHosts by Window
Summary:
We want to re-enable View recycling for ComponentHost as they're one of our most used Views. We had to disable it in the past due to getting a large amount of native RenderThread crashes - this is likely due to a bug in AOSP, but AOSP also doesn't expect this kind of recycling.

We have some evidence this may have to do with recycling Views across multiple Window instances, so this experiment is to try to scope recycling pools to a given Window instance.

To do this, we use the Window token as a key indicating the current window. To do this, we use a weak ref to it as a key to the recycling pool in addition to context. We are only implementing this for ComponentHost to test the hypothesis that this would give us recycling that avoids the native crash.

Reviewed By: pasqualeanatriello

Differential Revision: D33941763

fbshipit-source-id: 7210f430096d1ff7601d95212334400382583029