Avoid processing visibility outputs in RCMS when the flag for this behaviour is disabled.
Summary:
When notifyVisibleBoundsChanged is called with processVisibilityOutputs is false, we should not be processing the visibility outputs in VisibilityMountExtension. This is already the case when using Litho's MountState, but in RenderCore MountState flows, this flag is ignored, thereby processing visibility outputs always. Processing visibility outputs is sometimes not needed (such as during transitions), and can lead to redundant visibility events being invoked, and reducing overall perf.

In this diff, we leverage VisibilityMountExtension's input (LayoutState) to prevent the visibility extension from processing visibility outputs (in after-mount and onVisibleBoundsChanged) when the flag is false.

Reviewed By: adityasharat

Differential Revision: D33892531

fbshipit-source-id: 3f880c8e166c8ce8248984439199d0f22adcd785