Avoid recreating RenderTree in LayoutState toRenderTree
Summary:
To avoid recreating the render-tree when it's already been created, LayoutState will cache the first tree created in toRenderTree, then return that cached result in subsequent calls.
This change also means that RCMS can return early when mounting when checking if the RenderTree instance is the same as what it already has.

Reviewed By: astreet

Differential Revision: D34042968

fbshipit-source-id: 0cacf0bbed994c143a06f507114c28e11131780a