Delete bind / unbind static helper methods in RenderUnit in order to reduce stack frames
Summary: The bind / unbind helper methods in RenderUnit simply check the provided list is not null, then iterates the list. Due to the simplicity of these methods, we can remove them, and add the iterations in place where needed, thus avoiding an additional stack frame at the cost of some minor code duplication.

Reviewed By: astreet

Differential Revision: D34076676

fbshipit-source-id: b845f7978e3a3605f6606b0bd1b4d5c51a88d6e3