Adapt VisibilityEventsIncrementalMountDisabledTest tests for LithoView self-management.
Summary: Main issues here are due to the parent's sizes being defined in the test, and are too small to contain some layouts in various tests, causing the "corrected" visible logic to crop them out. To resolve this, the relevant layouts have been shrunk to fit inside the parent. 1 test is now being ignored when the litho-self-management config is on.

Reviewed By: mihaelao

Differential Revision: D33549040

fbshipit-source-id: ea133d0ed2076c4474296532e7090f42a211f410