Implement alternative to notifyVisibleBoundsChanged call via use of ViewTreeObserver listeners in LithoView
Summary:
Currently, LithoViews rely on external calls to notifyVisibleBoundsChanged when the LithoView's view port has potentially changed. This is problematic for the following reasons:
1. Ensuring notifyVisibleBoundsChanged is called whenever a LithoView's view port has changed means a lot of extra code complexity
2. Sometimes it is not possible to make this call, for example - when the LithoView is nested within a custom view. This problem leads to Incremental Mount not working.

To solve this, we introduce a mechanism that leverages ViewTreeObserver's OnScrollChangedListener & OnGlobalLayoutListener so that LithoViews can essentially self-manage their view-port change. When the callbacks are invoked, each LithoView will assess if their view-port has changed, and then trigger the notifyVisibleBoundsChanged sequence.

These solutions work for any layout or scroll changes but do not address animations or translation. To solve this, a custom implementation for getting the visible rect was introduced: getCorrectedLocalVisibleRect. The implementation is adapted from the android source-code, only without the addition of applying matrix transformations.

To properly test this behaviour, the previous behaviour has been left intact, and controlled via a new mobile config.

Reviewed By: mihaelao

Differential Revision: D33428861

fbshipit-source-id: 1b9251f4122e948b6da1488a916cbed2f68c7bd1