Move Extension into ExtensionState to avoid Map lookups
Summary:
**TL;DR;**
This is a performance optimisation. The goal of the diff is to avoid looking up in a map for every MountExtensions state for every lifecycle method. Since there is an 1:1 mapping from extension to its state. This diff just moves the Extension (which are stateless) into its `ExtensionState`.

**Details:**
Methods like `onMountItem`, `onUnmountItem`, `onBindItem`, `onUnbindItem`, will be invoked for every `RenderUnit`, and for each one of them every MountExtensions equivalent lifecycle method will be invoked. For each lifecycle call of every `MountExtension` a map is looked up to get its `ExtensionState`.

This diff adds the `MountExtension` into its `ExtensionState` to avoid map lookups altogether. As a bonus this improves code readability.

Files with interesting changes
* LithoHostListenerCoordinator.java
* ExtensionState.java
* MountDelegate.java

Differential Revision: D34042815

fbshipit-source-id: bdd46b0ddf1916cb9d6db627f13685c475507fce