Replaces usage of MountType with RenderType [2/n]
Summary:
The goal of this stack is to be able to add a new MountType which can used exclusively by MountableComponent to switch code flow when resolving the tree.

This will be done by replacing the use of MountType with RC's RenderType or simply by simple using an instanceof check where possible.

No functional changes.

Differential Revision: D34170727

fbshipit-source-id: d5997fd708b9b91f82dab9e4ac180506ea8133c3