Replaces usage of MountType with RenderType [3/n] and adds Mountable
Summary:
The goal of this stack is to be able to add a new MountType which can used exclusively by MountableComponent to switch code flow when resolving the tree.

This will be done by replacing the use of MountType with RC's RenderType or simply by simple using an instanceof check where possible.

This diff is also a bit different:
The code branches before creating the RenderUnit, and its implementation depends on the RenderType, therefore the `Mountable` is required to resolve that information.

This diff adds a new `Mountable` class i-var to `LithoLayoutResult`.

No functional changes.

Differential Revision: D34170725

fbshipit-source-id: 6a81be171afdbf86a2c31d9ab9832d627278fddf