Remove widthSpec and heightSpec from ComponentContext
Summary:
Width and Height spec removed from ComponentContext to make it lighter.

Component's render api will now accept width and height spec to pass them to OnCreateLayoutWithSizeSpec.
We are passing valid width and height specs only when we need to resolve nested tree because that's when we actually need them.

Reviewed By: adityasharat

Differential Revision: D33894876

fbshipit-source-id: 799e018b671ec564feeba18391f41e16c497a5a6