Litho Kotlin Coroutines 1: ComponentTree.componentTreeScope
Summary:
Running coroutines from KComponents requires a CoroutineScope to launch coroutines in. This diff adds a `componentTreeScope` that can be lazily created for a ComponentTree.

The main reason for this idea is that all the coroutine helper methods could just use this CoroutineScope directly under the hood, and so each KComponent wouldn't have to think about getting their own CoroutineScope to work within.

This componentTreeScope will be lazily created when it's first accessed, but will only remain active if and while the ComponentTree is not released. Once the ComponentTree is released, scope is canceled.

Reviewed By: acejingbo

Differential Revision: D34125835

fbshipit-source-id: 3063ad205f4293914169426976166f73dc1f57b1