Litho Kotlin Coroutines 4: useFlow
Summary:
This combines the concepts of `useState` and `Flow.collect`. This will collect from any Flow and update the State with any values it receives.

I think this is a nice way for a ViewModel to be able to expose observable state that should cause a KComponent to re-render itself.

Because flows are cold, there's often no default value, so an `initialValue` block is available. A variation for StateFlow will automatically use the current value of as the initialValue.

There are a few variations of `useFlow` available here:
1) Takes a single flow as an argument
2) Has a block that can return a new flow. Helpful if the flow should change based on some key dependencies.
3) Takes a StateFlow, so it can automatically detect the initial value.

Reviewed By: astreet

Differential Revision: D34055917

fbshipit-source-id: c92b7340191074f1b289de0dfd3d018d85243fcb