Litho Kotlin Coroutines 3: useProducer
Summary:
`useProducer` launches a coroutine whose purpose is to update a state value. It combines `useState` with something like `produce` in Kotlin except instead of sending values to a Channel, it updates a State.

This allows any coroutine to run and update some Litho `State` value, causing a component to re-render. Only the resulting value is available to the component. The update methods are only available to the coroutine.

This will be also be used for upcoming `useFlow`, where the producer coroutine collects a flow to produce state.

Reviewed By: astreet

Differential Revision: D34055918

fbshipit-source-id: b3f8a641509031e353158e034c3ecf799690ff3d