Make useCallback behave correctly
Summary:
The previous implementation of useCallback was quite broken: this diff fixes it as well as updates the tests to ensure that we are re-rendering only the expected number of times.

The approach is for us to have useCallback return a function of the same signature passed into useCallback, but which is a cached object that will maintain referential equality across render passes.

As Kotlin doesn't support decorators, I've had to do this by providing overloads for different function signatures, up to 7 params - from what I can tell, this is a pretty standard solution to this problem in Kotlin.

Reviewed By: andrewpmsmith

Differential Revision: D34178656

fbshipit-source-id: c6088d0302b9fb0da221a5385a76a48b99ad1432