Adds support for @PropDefault for Kotlin class spec
Summary:
Adds support for prop default in Kotlin class.

Static final members in a Kotlin class are compiled into an inner class called 'Companion'; because of this Litho's Annotation Processor doesn't find a 'default value' for any props. This diff fixes the problem by searching in a Companion class if present.

Reviewed By: astreet

Differential Revision: D34308176

fbshipit-source-id: 14559803df828f002021c5184f828faabb4bb31b