Fix Kotlin @PropDefault Only Extraction
Summary:
This fixes the case of:
```
PropDefault val justPropDefault: Long = 12345L
PropDefault val justPropDefault2: List<String> = listOf("Prop", "Default")
```

This case brings up a 3rd (and final?) case where we were not properly handling `PropDefault`:
| Type | Spec | Decompiled |
| -- | -- |
| Object | {F703765992} | {F703766100} |
| Class | {F703765681} | {F703765883} |

(The important part to see from the class spec is that -- the method we want to search for is within the `Companion` whereas previous fields are in the `top-level` spec.  This is addressed in the things of note section.)

The basic idea here is that--
1. We have the annotations method
2. The field now is *private* so we
3. Need to find the non-annotated getter method.

Therefore, we now have 3 different ways to access a Kotlin PropDefault:
1. Annotated getter method
2. Unannotated getter method (this diff!)
3. Field (prior diff!)

```
// 1. Annotated getter method
PropDefault
public int getPropDefault1() { return ... )

// 2. Unannotated getter method
PropDefault
public void getPropDefault2$annotations() {}

public int getPropDefault2() { return ... }

// 3. Field
public static int propDefault3 = 100

PropDefault
public void getPropDefault3$annotations() {}
```

## Things of Note
When setting up the `Class` spec, I found that we were also missing iterating elements for this type of `PropDefault` setup.  This is because we were limiting the search of elements and needed to include elements within the top level class *and* elements within the companion object.

Lastly, if you do:
```
1. enclosedElement = enclosingElement.getEnclosedElements().first()
2. elementsToSearch = enclosedElements.getEnclosingElement().getEnclosedElements()
```

That says:
- From Element A, get the first enclosed element, eElement1
- From eElement 1, get the enclosing element (this returns Element A)
- From Element A, get the enclosed elements...

That's a very roundabout way of saying... elementA.getEnclosedElements()

Therefore, I simplified things.

Reviewed By: adityasharat

Differential Revision: D34398526

fbshipit-source-id: b28572f255d55380b09cfec1c6f574d4cd389e40