Fix Kotlin Const & JvmField PropDefaults
Summary:
This diff adds support for the following type of `PropDefault` annotations:

```
PropDefault const val constProp: String = "Let's try this..."
PropDefault JvmField val fieldProp: List<String> = listOf("Let's try this'")
```

In a following diff, support will be added for:
```
PropDefault val fieldProp: List<String> = listOf("Let's try this'")
```

IMPORTANT: This now breaks `field:PropDefault` when building a spec that does that because the field is private (see prior diff).  (I'm not sure why worked in the first place.)  Therefore, once this stack of diffs land, I think I need to mark this as a breaking change.
------------------------------
# Problem

If you write your prop defaults like so, you'll never actually apply them:

```
PropDefault const val constProp: String = "Let's try this..."
PropDefault JvmField val fieldProp: List<String> = listOf("Let's try this'")
```

Instead, you *have* to write them like this:
```
get:PropDefault val propWithDefault: Int = Color.RED
```

Really, any of the above 3 *should work*, especially since the fix within FB right now is to lint if people do the first 2.  (Outside of Fb, there's a pretty good chance that things are just silently broken.)

# Solution

Essentially, though, this is what we're working with when we decompile the code:

```
NotNull
   public static final String constProp = "Let's try this...";
   JvmField
   NotNull
   public static final List fieldProp;

...

/** deprecated */
   // $FF: synthetic method
   PropDefault
   public static void getConstProp$annotations() {
   }

   /** deprecated */
   // $FF: synthetic method
   PropDefault
   public static void getFieldProp$annotations() {
   }
```

When we have a deprecated, synthetic method, we want to find the public element and directly access hat static field.  We can't use the void method.

Therefore, once we find a `get<PropName>$annotations()` method, we want to find and work with the public static constant that's not annotated.

Okay, once we have the prop default and are going to actually initialize it, we need to know:

1. Is this a case where we need to call a get method?
2. If Yes --> Is it a boolean so we drop the `get` or do we keep it?
3. If no, then we have a field accessor and want to access via the prop name.

This diff combines all of that together to at least get us the properly generated code.

## Class Spec: Edge Case

The case where it's within a class's companion object then we need to handle things differently.  Basically, this is because this is what the decompiled code looks like...

```
public final class KotlinClassSpec {
   private static final int propHasDefault = -65536;
   NotNull
   public static final String constProp = "Let's try this...";
   JvmField
   NotNull
   public static final List fieldProp = CollectionsKt.listOf("Let's try this'");
   NotNull
   public static final KotlinClassSpec.Companion Companion = new KotlinClassSpec.Companion((DefaultConstructorMarker)null);

   JvmStatic
   OnCreateLayout
   NotNull
   public static final Component onCreateLayout(NotNull ComponentContext c, Prop NotNull String data, Prop(optional = true) int propHasDefault, Prop(optional = true) NotNull String constProp, Prop(optional = true) NotNull List fieldProp) {
      return Companion.onCreateLayout(c, data, propHasDefault, constProp, fieldProp);
   }

   Metadata(
      mv = {1, 5, 1},
      k = 1,
      d1 = {"\u00008\n\u0002\u0018\u0002\n\u0002\u0010\u0000\n\u0002\b\u0002\n\u0002\u0010\u000e\n\u0002\b\u0002\n\u0002\u0010 \n\u0002\b\u0002\n\u0002\u0010\b\n\u0002\b\u0003\n\u0002\u0018\u0002\n\u0000\n\u0002\u0010\r\n\u0000\n\u0002\u0018\u0002\n\u0002\b\u0002\b\u0086\u0003\u0018\u00002\u00020\u0001B\u0007\b\u0002¢\u0006\u0002\u0010\u0002JH\u0010\r\u001a\u00020\u000e\"\b\b\u0000\u0010\u000f*\u00020\u00102\u0006\u0010\u0011\u001a\u00020\u00122\b\b\u0001\u0010\u0013\u001a\u00020\u00042\b\b\u0001\u0010\t\u001a\u00020\n2\b\b\u0001\u0010\u0003\u001a\u00020\u00042\u000e\b\u0001\u0010\u0006\u001a\b\u0012\u0004\u0012\u00020\u00040\u0007H\u0007R\u0016\u0010\u0003\u001a\u00020\u00048\u0006X\u0087T¢\u0006\b\n\u0000\u0012\u0004\b\u0005\u0010\u0002R\u001c\u0010\u0006\u001a\b\u0012\u0004\u0012\u00020\u00040\u00078\u0006X\u0087\u0004¢\u0006\b\n\u0000\u0012\u0004\b\b\u0010\u0002R\u0016\u0010\t\u001a\u00020\n8GX\u0086D¢\u0006\b\n\u0000\u001a\u0004\b\u000b\u0010\f¨\u0006\u0014"},
      d2 = {"Lcom/facebook/samples/litho/KotlinClassSpec$Companion;", "", "()V", "constProp", "", "getConstProp$annotations", "fieldProp", "", "getFieldProp$annotations", "propHasDefault", "", "getPropHasDefault", "()I", "onCreateLayout", "Lcom/facebook/litho/Component;", "T", "", "c", "Lcom/facebook/litho/ComponentContext;", "data", "fbandroid_libraries_components_sample_src_main_java_com_facebook_samples_litho"}
   )
   public static final class Companion {
      PropDefault
      public final int getPropHasDefault() {
         return KotlinClassSpec.propHasDefault;
      }

      /** deprecated */
      // $FF: synthetic method
      PropDefault
      public static void getConstProp$annotations() {
      }

      /** deprecated */
      // $FF: synthetic method
      PropDefault
      public static void getFieldProp$annotations() {
      }

      JvmStatic
      OnCreateLayout
      NotNull
      public final Component onCreateLayout(NotNull ComponentContext c, Prop NotNull String data, Prop(optional = true) int propHasDefault, Prop(optional = true) NotNull String constProp, Prop(optional = true) NotNull List fieldProp) {
         Intrinsics.checkNotNullParameter(c, "c");
         Intrinsics.checkNotNullParameter(data, "data");
         Intrinsics.checkNotNullParameter(constProp, "constProp");
         Intrinsics.checkNotNullParameter(fieldProp, "fieldProp");
         Column var10000 = Column.create(c).build();
         Intrinsics.checkNotNullExpressionValue(var10000, "Column.create(c).build()");
         return (Component)var10000;
      }

      private Companion() {
      }

      // $FF: synthetic method
      public Companion(DefaultConstructorMarker $constructor_marker) {
         this();
      }
   }
}
```

The fields are static on the class and not on the companion.  So, we need to access them there instead of on the `Companion`.

-------------------------

Reviewed By: adityasharat

Differential Revision: D34331629

fbshipit-source-id: bca8b21dcca9205d8d53456396dfa681c3ba652e