Expose LazyCollectionController to access RecyclerView. Remove RecyclerEventsController
Summary:
We had to expose `RecyclerEventsController` on `LazyCollection` as it is being used by EventsPermalink. We tried to dissuade people from using `RecyclerEventsController` or `RecyclerCollectionEventsController` with a comment. But it wasn't enforced.

This diff stops exposing recyclerEventsController, and replace with LazyCollectionController. This gives a way to expose the RecyclerView (used by EventPermalink), without exposing all the other features like scrolling, which we don't want being used.

If we need to expose more things, we can copy them over case-by-case. This way we have more control.

Reviewed By: astreet

Differential Revision: D34276936

fbshipit-source-id: b864b405372bef38115d7ff41018d6a88a1f8f3d